requirements.txt
playwright~=1.52.0
allure-python-commons~=2.14.2
pytest~=8.3.5
Faker~=37.1.0
python-dotenv~=1.1.0
requests~=2.32.3